Refactored utils into browser_utils and checks_utils
