+++ checks/browsers/chromium/searchEngine_chromium.go
@@ -3,6 +3,7 @@ package chromium
 import (
 	"encoding/json"
 	"fmt"
+	browser_utils "github.com/InfoSec-Agent/InfoSec-Agent/checks/browsers"
 	"io"
 	"log"
 	"os"
@@ -12,7 +13,6 @@ import (
 	"github.com/InfoSec-Agent/InfoSec-Agent/mocking"
 
 	"github.com/InfoSec-Agent/InfoSec-Agent/checks"
-	"github.com/InfoSec-Agent/InfoSec-Agent/utils"
 )
 
 // SearchEngineChromium inspects the default search engine setting in Chromium-based browsers.
@@ -55,7 +55,7 @@ func SearchEngineChromium(browser string) checks.Check {
 	}
 	defer func(file *os.File) {
 		tmpFile := mocking.Wrap(file)
-		err = utils.CloseFile(tmpFile)
+		err = browser_utils.CloseFile(tmpFile)
 		if err != nil {
 			log.Println("Error closing file")
 		}
