+++ checks/browsers/firefox/history_firefox.go
@@ -2,14 +2,13 @@ package firefox
 
 import (
 	"database/sql"
+	browser_utils "github.com/InfoSec-Agent/InfoSec-Agent/checks/browsers"
 	"os"
 	"path/filepath"
 	"regexp"
 	"time"
 
 	"github.com/InfoSec-Agent/InfoSec-Agent/logger"
-
-	"github.com/InfoSec-Agent/InfoSec-Agent/utils"
 
 	"github.com/InfoSec-Agent/InfoSec-Agent/checks"
 	// Necessary to use the sqlite driver
@@ -26,7 +25,7 @@ import (
 // This function works by locating the Firefox profile directory and copying the places.sqlite database to a temporary location. It then opens this database and queries it for the URLs visited in the last week. It processes the results of the query by checking each URL against a list of known phishing domains. If a match is found, a string is generated that includes the domain name and the time of the visit, and this string is added to the results. If any error occurs during this process, such as an error copying the file or querying the database, this error is returned as the result of the check.
 func HistoryFirefox() checks.Check {
 	var output []string
-	ffdirectory, err := utils.FirefoxFolder()
+	ffdirectory, err := browser_utils.FirefoxFolder()
 	if err != nil {
 		return checks.NewCheckErrorf(checks.HistoryFirefoxID, "No firefox directory found", err)
 	}
@@ -42,7 +41,7 @@ func HistoryFirefox() checks.Check {
 	}(tempHistoryDbff)
 
 	// Copy the database to a temporary location
-	copyError := utils.CopyFile(ffdirectory[0]+"\\places.sqlite", tempHistoryDbff, nil, nil)
+	copyError := browser_utils.CopyFile(ffdirectory[0]+"\\places.sqlite", tempHistoryDbff, nil, nil)
 	if copyError != nil {
 		return checks.NewCheckError(checks.HistoryFirefoxID, copyError)
 	}
@@ -133,7 +132,7 @@ func closeRows(rows *sql.Rows) {
 // This function iterates over each row in the provided result set. For each row, it extracts the URL and last visit date, and checks the URL against a list of known phishing domains. If a match is found, a string is generated that includes the domain name and the time of the visit, and this string is added to the results. If an error occurs during this process, it is returned as well.
 func processQueryResults(rows *sql.Rows) ([]string, error) {
 	var output []string
-	phishingDomainList := utils.GetPhishingDomains()
+	phishingDomainList := browser_utils.GetPhishingDomains()
 	re := regexp.MustCompile(`(?:https?:\/\/)?(?:[^@\n]+@)?(?:www\.)?([^:\/\n?]+\.[^:\/\n?]+)`)
 
 	for rows.Next() {
