+++ checks/browsers/firefox/password_firefox.go
@@ -2,13 +2,13 @@ package firefox
 
 import (
 	"encoding/json"
+	browser_utils "github.com/InfoSec-Agent/InfoSec-Agent/checks/browsers"
 	"os"
 	"strconv"
 
 	"github.com/InfoSec-Agent/InfoSec-Agent/logger"
 
 	"github.com/InfoSec-Agent/InfoSec-Agent/checks"
-	"github.com/InfoSec-Agent/InfoSec-Agent/utils"
 	"github.com/andrewarchi/browser/firefox"
 )
 
@@ -22,7 +22,7 @@ import (
 // This function first determines the directory in which the Firefox profile is stored. It then opens the 'logins.json' file, which contains a list of all saved Firefox passwords. The function decodes the JSON file into a struct, and then iterates over the struct to extract the saved passwords. These passwords are added to the results, which are returned as a Check object. If an error occurs at any point during this process, it is encapsulated in the Check object and returned.
 func PasswordFirefox() checks.Check {
 	// Determine the directory in which the Firefox profile is stored
-	ffdirectory, _ := utils.FirefoxFolder()
+	ffdirectory, _ := browser_utils.FirefoxFolder()
 
 	var output []string
 	// Open the logins.json file, which contains a list of all saved Firefox passwords
